+++ pysmashgg/p_queries.py
@@ -71,3 +71,121 @@ PLAYER_SHOW_TOURNAMENTS_FOR_GAME_QUERY = """query ($playerId: ID!, $playerName: 
     }
   }
 }"""
+
+PLAYER_BY_SLUG_QUERY = """query Profile($discriminatorSlug: String!) {
+  user(slug: $discriminatorSlug) {
+    id
+    name
+    player {
+      id
+      gamerTag
+      prefix
+    }
+    bio
+    birthday
+    discriminator
+    slug
+    location {
+      country
+      state
+      city
+    }
+    authorizations(types: [TWITTER, TWITCH, DISCORD]) {
+      type
+      externalUsername
+      url
+    }
+    images {
+      id
+      url
+      height
+      width
+      ratio
+      type
+    }
+    events(query:{page:1,perPage:10}) {
+      nodes {
+        id
+        name
+        numEntrants
+        slug
+        startAt
+        videogame {
+          id
+          name
+        }
+      }
+    }
+  }
+}"""
+
+PLAYER_RECENT_PLACEMENTS_QUERY = """query ($playerId: ID!, $videogameId: ID!) {
+  player(id: $playerId) {
+    id
+    gamerTag
+    prefix
+    user {
+      name
+      discriminator
+      slug
+      images {
+        height
+        width
+        url
+        type
+      }
+      authorizations(types: [TWITTER, TWITCH, DISCORD]) {
+        type
+        externalUsername
+        url
+      }
+      location {
+        country
+        state
+        city
+      }
+    }
+    recentStandings(videogameId: $videogameId, limit: 20) {
+      id
+      placement
+      entrant {
+        id
+        name
+        event {
+          id
+          name
+          slug
+          isOnline
+          numEntrants
+          tournament {
+            name
+            startAt
+            images {
+              url
+              width
+              height
+              type
+            }
+          }
+          videogame {
+            displayName
+            images {
+              type
+              url
+              width
+              height
+            }
+          }
+        }
+      }
+    }
+  }
+}"""
+
+PLAYER_LOOKUP_ID_QUERY = """query LookupPlayerId($discriminatorSlug: String!) {
+  user(slug: $discriminatorSlug) {
+    player {
+      id
+    }
+  }
+}"""